# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rekognition Manager BFF API",
    "description": "Amazon Rekognitionのコレクション、ユーザー、顔を管理するBFF API",
    "version": "0.7.0"
  },
  "servers": [{ "url": "http://localhost:3001", "description": "ローカルBFF" }],
...
  "paths": {
    "/api/v1/collections": {
      "get": {
        "operationId": "listCollections",
        "summary": "コレクション一覧を取得する",
        "tags": ["Collections"],
        "responses": {
          "200": {
            "description": "コレクション一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "collectionId": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 255,
                            "pattern": "^[a-zA-Z0-9_.-]+$"
                          },
                          "collectionArn": { "type": "string" },
                          "faceModelVersion": { "type": "string" },
                          "creationTimestamp": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          },
                          "faceCount": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "userCount": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": ["collectionId"]
                      }
                    },
                    "nextCursor": { "anyOf": [{ "type": "string" }, { "type": "null" }] }
                  },
                  "required": ["items", "nextCursor"]
                }
              }
            }
          },
          "400": {
            "description": "入力エラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": { "type": "string" },
                              "message": { "type": "string" }
                            },
                            "required": ["path", "message"]
                          }
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "対象が見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": { "type": "string" },
                              "message": { "type": "string" }
                            },
                            "required": ["path", "message"]
                          }
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "サーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": { "type": "string" },
                              "message": { "type": "string" }
                            },
                            "required": ["path", "message"]
                          }
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "parameters": [
          { "in": "query", "name": "cursor", "schema": { "type": "string", "minLength": 1 } },
          {
            "in": "query",
            "name": "limit",
            "schema": { "default": 50, "type": "integer", "minimum": 1, "maximum": 100 }
          }
        ]
      },
      "post": {
        "operationId": "createCollection",
        "summary": "コレクションを作成する",
        "tags": ["Collections"],
        "responses": {
          "201": {
            "description": "作成したコレクション",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "collectionId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255,
                      "pattern": "^[a-zA-Z0-9_.-]+$"
                    },
                    "collectionArn": { "type": "string" },
                    "faceModelVersion": { "type": "string" },
                    "creationTimestamp": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "faceCount": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 },
                    "userCount": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 }
                  },
                  "required": ["collectionId"]
                }
              }
            }
          },
          "400": {
            "description": "入力エラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": { "type": "string" },
                              "message": { "type": "string" }
                            },
                            "required": ["path", "message"]
                          }
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "対象が見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": { "type": "string" },
                              "message": { "type": "string" }
                            },
                            "required": ["path", "message"]
                          }
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "サーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": { "type": "string" },
                              "message": { "type": "string" }
                            },
                            "required": ["path", "message"]
                          }
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "collectionId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "pattern": "^[a-zA-Z0-9_.-]+$"
                  }
                },
                "required": ["collectionId"]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/collections/{collectionId}/users/{userId}/faces": {
      "post": {
        "operationId": "associateFaces",
        "summary": "顔をユーザーへ関連付ける",
        "tags": ["Users", "Faces"],
        "responses": {
          "200": {
            "description": "関連付け結果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "associatedFaceIds": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                      }
                    },
                    "unsuccessful": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "faceId": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                          },
                          "userId": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 128,
                            "pattern": "^[a-zA-Z0-9_.:-]+$"
                          },
                          "confidence": { "type": "number" },
                          "reasons": { "type": "array", "items": { "type": "string" } }
                        },
                        "required": ["reasons"]
                      }
                    },
                    "userStatus": {
                      "type": "string",
                      "enum": ["ACTIVE", "UPDATING", "CREATING", "CREATED"]
                    }
                  },
                  "required": ["associatedFaceIds", "unsuccessful"]
                }
              }
            }
          },
          "400": {
            "description": "入力エラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": { "type": "string" },
                              "message": { "type": "string" }
                            },
                            "required": ["path", "message"]
                          }
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "対象が見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": { "type": "string" },
                              "message": { "type": "string" }
                            },
                            "required": ["path", "message"]
                          }
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "サーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "requestId": { "type": "string" },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": { "type": "string" },
                              "message": { "type": "string" }
                            },
                            "required": ["path", "message"]
                          }
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "collectionId",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "pattern": "^[a-zA-Z0-9_.-]+$"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[a-zA-Z0-9_.:-]+$"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "faceIds": {
                    "minItems": 1,
                    "maxItems": 100,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                    }
                  },
                  "userMatchThreshold": { "type": "number", "minimum": 0, "maximum": 100 }
                },
                "required": ["faceIds"]
              }
            }
          }
        }
      }
    },
...
  },
...
}